# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentForge API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/agents": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "List Agents",
        "operationId": "list_agents_api_v1_agents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentResponse"
                  },
                  "type": "array",
                  "title": "Response List Agents Api V1 Agents Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "Create Agent",
        "operationId": "create_agent_api_v1_agents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/knowledge/sources": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "List Sources",
        "operationId": "list_sources_api_v1_knowledge_sources_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/KnowledgeSourceOut"
                  },
                  "type": "array",
                  "title": "Response List Sources Api V1 Knowledge Sources Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/knowledge/ingest": {
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Ingest",
        "operationId": "ingest_api_v1_knowledge_ingest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeIngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeIngestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/knowledge/upload": {
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Upload",
        "operationId": "upload_api_v1_knowledge_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_api_v1_knowledge_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeIngestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/knowledge/sources/{title}": {
      "delete": {
        "tags": [
          "knowledge"
        ],
        "summary": "Delete Source",
        "operationId": "delete_source_api_v1_knowledge_sources__title__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "title",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Title"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/webhooks": {
      "get": {
        "tags": [
          "webhooks"
        ],
        "summary": "List Webhooks",
        "operationId": "list_webhooks_api_v1_webhooks_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WebhookResponse"
                  },
                  "type": "array",
                  "title": "Response List Webhooks Api V1 Webhooks Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "webhooks"
        ],
        "summary": "Create Webhook",
        "operationId": "create_webhook_api_v1_webhooks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/webhooks/{webhook_id}": {
      "delete": {
        "tags": [
          "webhooks"
        ],
        "summary": "Delete Webhook",
        "operationId": "delete_webhook_api_v1_webhooks__webhook_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "webhook_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Webhook Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "graph_definition": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GraphDefinitionValidated"
              },
              {
                "additionalProperties": true,
                "type": "object"
              }
            ],
            "title": "Graph Definition"
          },
          "model_config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentModelConfig"
              },
              {
                "additionalProperties": true,
                "type": "object"
              }
            ],
            "title": "Model Config"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills",
            "description": "Registry skill UUIDs; tool nodes use config.tool_name equal to skill.name. Built-in tools: echo, fetch, retrieve (RAG over /api/v1/knowledge/ingest)."
          },
          "execution_policy": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Policy"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "AgentCreateRequest"
      },
...
      "AgentModelConfig": {
        "properties": {
          "provider": {
            "type": "string",
            "enum": [
              "mock",
              "openai",
              "google",
              "gemini",
              "finetuned",
              "anthropic"
            ],
            "title": "Provider",
            "default": "mock"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "finetune_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finetune Job Id"
          }
        },
        "type": "object",
        "title": "AgentModelConfig"
      },
      "AgentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "graph_definition": {
            "title": "Graph Definition"
          },
          "model_config": {
            "title": "Model Config"
          },
          "interrupt_config": {
            "title": "Interrupt Config"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "security_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Security Score"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "name",
          "description",
          "graph_definition",
          "model_config",
          "interrupt_config",
          "skills",
          "status",
          "security_score",
          "created_at",
          "updated_at"
        ],
        "title": "AgentResponse"
      },
...
      "Body_upload_api_v1_knowledge_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_api_v1_knowledge_upload_post"
      },
...
      "GraphDefinitionValidated": {
        "properties": {
          "graph_schema_version": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Graph Schema Version",
            "description": "AgentForge Graph (AFG) schema revision; JSON in DB is canonical.",
            "default": "1.0"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNode"
            },
            "type": "array",
            "minItems": 1,
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdge"
            },
            "type": "array",
            "title": "Edges"
          },
          "entry_point": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Entry Point"
          },
          "parallel_nodes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Parallel Nodes",
            "description": "Optional hint: node ids intended for parallel execution (orchestrator-specific)."
          }
        },
        "type": "object",
        "required": [
          "nodes",
          "entry_point"
        ],
        "title": "GraphDefinitionValidated",
        "description": "§6.1 shape: nodes, edges, entry_point."
      },
      "GraphEdge": {
        "properties": {
          "from": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "From"
          },
          "to": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "To"
          },
          "condition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Condition"
          },
          "condition_type": {
            "type": "string",
            "enum": [
              "contains",
              "regex",
              "json_path",
              "always"
            ],
            "title": "Condition Type",
            "default": "contains"
          }
        },
        "type": "object",
        "required": [
          "from",
          "to"
        ],
        "title": "GraphEdge"
      },
      "GraphNode": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Id"
          },
          "type": {
            "type": "string",
            "maxLength": 64,
            "title": "Type",
            "description": "Built-in or plugin-registered node type.",
            "default": "llm"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "GraphNode"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KnowledgeIngestRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Title"
          },
          "text": {
            "type": "string",
            "maxLength": 500000,
            "minLength": 1,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "title",
          "text"
        ],
        "title": "KnowledgeIngestRequest"
      },
      "KnowledgeIngestResponse": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "chunks": {
            "type": "integer",
            "title": "Chunks"
          }
        },
        "type": "object",
        "required": [
          "title",
          "chunks"
        ],
        "title": "KnowledgeIngestResponse"
      },
      "KnowledgeSourceOut": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count"
          }
        },
        "type": "object",
        "required": [
          "title",
          "chunk_count"
        ],
        "title": "KnowledgeSourceOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WebhookCreateRequest": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url"
          },
          "events": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Events",
            "description": "Subset of: execution.completed, campaign.completed"
          },
          "secret": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Secret",
            "description": "Optional HMAC-SHA256 secret for X-AgentForge-Signature header"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "WebhookCreateRequest"
      },
      "WebhookResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "events": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Events"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "url",
          "events",
          "active",
          "created_at"
        ],
        "title": "WebhookResponse"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}